backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
trendspyg>=0.4.0
rank-bm25>=0.2.2
openai>=1.30.0
anthropic>=0.30.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
pydantic>=2.0.0
sse-starlette>=2.0.0
aiofiles>=24.0.0
pyyaml>=6.0.0
